+++ efootprint/builders/hardware/servers_boaviztapi.py
@@ -1,6 +1,6 @@
 import requests
 
-from efootprint.builders.hardware.storage_defaults import default_ssd
+from efootprint.builders.hardware.storage_defaults import default_ssd, default_hdd
 from efootprint.constants.sources import Source, Sources
 from efootprint.abstract_modeling_classes.source_objects import SourceValue
 from efootprint.core.hardware.servers.autoscaling import Autoscaling
@@ -107,7 +107,14 @@ def get_cloud_server(
     if base_cpu_consumption is None:
         base_cpu_consumption = SourceValue(0 * u.core, Sources.HYPOTHESIS)
     if storage is None:
-        storage = default_ssd()
+        reference_storage_capacity = SourceValue(1 * u.TB, source=Sources.HYPOTHESIS)
+        cloud_storage_capacity = SourceValue(32 * u.GB, source=Sources.HYPOTHESIS)
+        ratio = (cloud_storage_capacity / reference_storage_capacity).to(u.dimensionless).set_label(
+            "Ration of cloud server storage capacity to reference storage capacity")
+        storage = default_ssd(
+            carbon_footprint_fabrication=SourceValue(160 * u.kg, Sources.STORAGE_EMBODIED_CARBON_STUDY) * ratio,
+            power=SourceValue(1.3 * u.W, Sources.STORAGE_EMBODIED_CARBON_STUDY) * ratio,
+            storage_capacity=SourceValue(32 * u.GB, source=Sources.HYPOTHESIS))
 
     impact_url = "https://api.boavizta.org/v1/cloud/instance"
     params = {"provider": provider, "instance_type": instance_type}
@@ -161,6 +168,15 @@ def on_premise_server_from_config(
     cpu_spec = impact_data["verbose"]["CPU-1"]
     ram_spec = impact_data["verbose"]["RAM-1"]
 
+    storage_type = None
+    storage_spec = None
+    if "SSD-1" in impact_data["verbose"].keys():
+        storage_type = "SSD"
+        storage_spec = impact_data["verbose"]["SSD-1"]
+    elif "HDD-1" in impact_data["verbose"].keys():
+        storage_type = "HDD"
+        storage_spec = impact_data["verbose"]["HDD-1"]
+
     if lifespan is None:
         lifespan = SourceValue(6 * u.year, Sources.HYPOTHESIS)
     if idle_power is None:
@@ -174,7 +190,41 @@ def on_premise_server_from_config(
     if base_cpu_consumption is None:
         base_cpu_consumption = SourceValue(0 * u.core, Sources.HYPOTHESIS)
     if storage is None:
-        storage = default_ssd()
+        storage_unit = getattr(u, storage_spec["capacity"]["unit"])
+        reference_storage_capacity = SourceValue(1 * u.TB, source=Sources.HYPOTHESIS)
+        cloud_storage_capacity = SourceValue(
+                    storage_spec["capacity"]["value"] * storage_unit,
+                    source=impact_source
+                )
+        ratio = (cloud_storage_capacity / reference_storage_capacity).to(u.dimensionless).set_label(
+            "Ration of cloud server storage capacity to reference storage capacity")
+        nb_unit = SourceValue(storage_spec["units"]["value"] * u.dimensionless, impact_source)
+        if storage_type == 'SSD':
+            storage = default_ssd(
+                storage_capacity=cloud_storage_capacity,
+                fixed_nb_of_instances=SourceValue(
+                    storage_spec["units"]["value"] * u.dimensionless, source=impact_source
+                ),
+                carbon_footprint_fabrication=SourceValue(
+                    storage_spec["impacts"]["gwp"]["embedded"]["value"] * u.kg, source=impact_source
+                ) / nb_unit * ratio,
+                power=SourceValue(1.3 * u.W, Sources.STORAGE_EMBODIED_CARBON_STUDY) * ratio,
+            )
+        elif storage_type == 'HDD':
+            storage = default_hdd(
+                storage_capacity=cloud_storage_capacity,
+                fixed_nb_of_instances=SourceValue(
+                    storage_spec["units"]["value"] * u.dimensionless, source=impact_source
+                ),
+                carbon_footprint_fabrication=SourceValue(
+                    storage_spec["impacts"]["gwp"]["embedded"]["value"] * u.kg, source=impact_source
+                ) / nb_unit * ratio,
+                average_carbon_intensity=average_carbon_intensity,
+                power_usage_effectiveness=power_usage_effectiveness,
+                power=SourceValue(4.2 * u.W, Sources.STORAGE_EMBODIED_CARBON_STUDY) * ratio,
+            )
+        else:
+            raise ValueError(f"Storage type {storage_type} not yet implemented")
 
     average_power_value = impact_data["verbose"]["avg_power"]["value"]
     average_power_unit = impact_data["verbose"]["avg_power"]["unit"]
@@ -185,7 +235,9 @@ def on_premise_server_from_config(
 
     return OnPremise(
         name,
-        carbon_footprint_fabrication=SourceValue(impacts["gwp"]["embedded"]["value"] * u.kg, impact_source),
+        carbon_footprint_fabrication=SourceValue(
+            (impacts["gwp"]["embedded"]["value"] * u.kg-storage_spec["impacts"]["gwp"]["embedded"]["value"] * u.kg)
+            , impact_source),
         # TODO: document and challenge power calculation
         power=SourceValue(average_power_value * u.W, impact_source),
         lifespan=lifespan,