+++ efootprint/builders/hardware/storage_defaults.py
@@ -11,8 +11,6 @@ def default_ssd(name="Default SSD storage", **kwargs):
         "lifespan": SourceValue(6 * u.years, Sources.HYPOTHESIS),
         "idle_power": SourceValue(0 * u.W, Sources.HYPOTHESIS),
         "storage_capacity": SourceValue(1 * u.TB, Sources.STORAGE_EMBODIED_CARBON_STUDY),
-        "power_usage_effectiveness": SourceValue(1.2 * u.dimensionless, Sources.HYPOTHESIS),
-        "average_carbon_intensity": SourceValue(100 * u.g / u.kWh, Sources.HYPOTHESIS),
         "data_replication_factor": SourceValue(3 * u.dimensionless, Sources.HYPOTHESIS),
         "base_storage_need": SourceValue(0 * u.TB, Sources.HYPOTHESIS),
         "data_storage_duration": SourceValue(5 * u.year, Sources.HYPOTHESIS)
@@ -30,8 +28,6 @@ def default_hdd(name="Default HDD storage", **kwargs):
         "lifespan": SourceValue(4 * u.years, Sources.HYPOTHESIS),
         "idle_power": SourceValue(0 * u.W, Sources.HYPOTHESIS),
         "storage_capacity": SourceValue(1 * u.TB, Sources.STORAGE_EMBODIED_CARBON_STUDY),
-        "power_usage_effectiveness": SourceValue(1.2 * u.dimensionless, Sources.HYPOTHESIS),
-        "average_carbon_intensity": SourceValue(100 * u.g / u.kWh, Sources.HYPOTHESIS),
         "data_replication_factor": SourceValue(3 * u.dimensionless, Sources.HYPOTHESIS),
         "base_storage_need": SourceValue(0 * u.TB, Sources.HYPOTHESIS),
         "data_storage_duration": SourceValue(5 * u.year, Sources.HYPOTHESIS)